Ensuring apt_update can be run from mco. Adding apt_clear_broken
This allows us to run apt_update directly from the mcollective tool, and also allows us to run `apt_clean_broken`,
which is a new command added in this commit - It clears out all partially configured packages. The way APT works, if
there is a package that does not finish completely (Normally due to the .deb postinst script not exiting with a <0 exit
code) - If a fixed package is now available on the repository, even after an apt-get update and install of the package again,
the new package is not used - instead APT continues to use the old, broken package. Clearing out partially configured packages
makes it so you can once again install the package.