MCOP-18 - check more explicitly for the Windows Provider
In addressing MCOP-14 in commit 90da47a I made an assumption that
all package providers on windows would be the windows package
provider from the puppet core.  This is a poor assumption as there
is a chocolatey provider which behaves differently.

Here we more specifically test the class of the provider is
`Puppet::Type::Package::ProviderWindows`